refactor(core): Make the datetime decoding shareable between JSON and XML decoding
